colons (:) in keys are now converted to periods (.)
Keys that contain colons would count as variables even when present in
comments and trigger confusing "Number of bound variables, doesn't match
number of parameters" errors. For whatever reason this is not always the
case; to avoid the pesky error the keys are sanitized so the system will
accept them in all cases.