slugid definition for pdx routines
The slugid is meant to be a string id associated with an entry. String
id's are extremely slow even when very small so it's best to avoid them
however in code they are easier to work with and produce more readable
code; it's also best not to rely on ids as they can change.

See acctg for use cases.