key/lang support in prepared statements deprecated
Statement keys was used to identify statements.
Language was used to identify the SQL language.

Both added a lot of "fluff" to prepared statements, and while the key
feature added some utility, it definetly didnt add enough to be worth it.
Since other more subtle solutions to the problems key/lang would solve are
both easy to integrate and potentially more useful. As such the fluff as
of 3.x support for key/lang is compeltely deprecated with no backwards
compatibility support.